requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# Vector database
faiss-cpu==1.7.4
numpy==1.26.0

# OpenAI
openai==1.12.0
httpx>=0.27.0,<0.28 # Required for OpenAI client

# Utilities
python-multipart==0.0.6
jinja2==3.1.2
markdown==3.5
tiktoken==0.5.1


